requirements.txt
# ─── Web Framework ────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-dotenv==1.0.1
pydantic==2.9.0

# ─── Document Processing ──────────────────────────────────────────────────────
langchain==0.3.7
langchain-community==0.3.7
langchain-text-splitters==0.3.0
langchain-experimental==0.3.0      # SemanticChunker
langchain-huggingface==0.1.0       # HuggingFaceEmbeddings

# ─── PDF Extraction ───────────────────────────────────────────────────────────
pymupdf==1.24.10

# ─── Embeddings & Reranking ───────────────────────────────────────────────────
sentence-transformers==3.1.1
transformers==4.45.0

# ─── Vector Store ─────────────────────────────────────────────────────────────
faiss-cpu==1.12.0

# ─── Lexical Search ───────────────────────────────────────────────────────────
rank-bm25==0.2.2

# ─── LLM (OpenRouter — unified gateway to Gemini / GPT-4o / Claude etc.) ─────
requests>=2.31.0

# ─── Utilities ────────────────────────────────────────────────────────────────
numpy==1.26.4
tqdm==4.66.5
psutil==5.9.8
nltk==3.9.2
